Fix lazy mode
We removed the outgoing flow from the kernel when we closed the SA.
We would never re-add it, because flow_loaded was still set in the
policy.

While we can simply reset the field and have the flow re-inserted,
it would create a blackhole between the removal and a subsequent
re-addition. It's better to avoid removal altogether.

+++ iked/iked.h
@@ -139,7 +139,8 @@ struct iked_flow {
 	struct iked_addr		 flow_dst;
 	unsigned int			 flow_dir;	/* in/out */
 
-	unsigned int			 flow_loaded;	/* pfkey done */
+	/* flow_loaded must be signed! */
+	int8_t				 flow_loaded;	/* pfkey done */
 
 	uint8_t				 flow_saproto;
 	uint8_t				 flow_ipproto;
